# Add a faster siphash24 implementation than the reference one