--- a/src/siphash24.h
+++ b/src/siphash24.h
@@ -1,15 +1,13 @@
-
 #pragma once
 
 #define SIPHASH_KEYLEN 16
-#define SIPHASH_HASHLEN 8
 
 extern "C" {
-int _siphash(uint8_t *out, const uint8_t *in, uint64_t inlen, const uint8_t *k);
+	uint64_t siphash24(const void *src, unsigned long src_sz, const uint64_t* key);
 }
 
 // [Bro] Wrapper for better type-safety.
 inline void siphash(uint64_t* digest, const uint8_t *in, uint64_t inlen, const uint8_t* key)
 	{
-	_siphash((uint8_t*)digest, in, inlen, key);
+	*digest = siphash24(in, inlen, (const uint64_t*)key);
 	}