- Average of 10 runs of 2009-M57-day11-18.trace (release build at -O3):
  - Master: 6.027s  93650 bytes max RSS
  - Commit: 5.950s  93271 bytes max RSS

- Hashing a fixed 32-byte payload 10 million times with a fixed key:
  - Master: 1.397411s
  - Commit: 0.998211s